+++ docker-entrypoint.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 
 python manage.py migrate
+python manage.py setup
 python manage.py createsuperuser --noinput
 gunicorn -b 0.0.0.0:8000 config.wsgi